refactor: mais algumas organizações de pastas
