+++ src/AuthApi.Application/AuthApi.Application.csproj
@@ -15,7 +15,22 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\AuthApi.Shared\AuthApi.Shared.csproj" />
+    <ProjectReference Include="..\AuthApi.Infrastructure\AuthApi.Infrastructure.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Compile Update="Resource\AuthApi_Resource.Designer.cs">
+      <DesignTime>True</DesignTime>
+      <AutoGen>True</AutoGen>
+      <DependentUpon>AuthApi_Resource.resx</DependentUpon>
+    </Compile>
+  </ItemGroup>
+
+  <ItemGroup>
+    <EmbeddedResource Update="Resource\AuthApi_Resource.resx">
+      <Generator>PublicXFileCodeGenerator</Generator>
+      <LastGenOutput>AuthApi_Resource.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
   </ItemGroup>
 
 </Project>
