+++ src/AuthApi.Application/Resource/AuthApi_Resource.resx
