+++ src/AuthApi.WebApi/AuthApi.WebApi.csproj
@@ -14,6 +14,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\AuthApi.Application\AuthApi.Application.csproj" />
+    <ProjectReference Include="..\AuthApi.Infrastructure\AuthApi.Infrastructure.csproj" />
   </ItemGroup>
 
 </Project>
