+++ src/AuthApi.WebApi/IoC/Configurations.cs
@@ -2,11 +2,11 @@
 using AuthApi.Application.Features.Users.RegisterUser.v1;
 using Microsoft.AspNetCore.Mvc;
 
-namespace AuthApi.WebApi.Configurations;
+namespace AuthApi.WebApi.IoC;
 
 public static class Configurations
 {
-    public static IServiceCollection AddCustomConfigurations(this IServiceCollection services)
+    public static IServiceCollection AddConfigurations(this IServiceCollection services)
     {
         services.AddScoped<UserRepository>();
         services.AddScoped<RegisterUserHandler>();
