+++ src/auth-api.sln
@@ -7,7 +7,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests", "..\tests\Tests.csp
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AuthApi.Application", "AuthApi.Application\AuthApi.Application.csproj", "{3E25ABE5-80F3-4D59-87C2-638BE761B56E}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AuthApi.Shared", "AuthApi.Shared\AuthApi.Shared.csproj", "{44DE2E50-D7C5-4D54-B3D5-29499512CD70}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AuthApi.Infrastructure", "AuthApi.Infrastructure\AuthApi.Infrastructure.csproj", "{44DE2E50-D7C5-4D54-B3D5-29499512CD70}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AuthApi.WebApi", "AuthApi.WebApi\AuthApi.WebApi.csproj", "{7635327D-FF87-496A-B636-E16D4D73E84B}"
 EndProject
